api/requirements.txt
yt-dlp==2023.7.6
uvicorn[standard]
fastapi==0.101.0
pydantic==2.1.1
celery[redis]==5.3.1
python-socketio[client]==5.8.0